refactored http-agent creation code
Added all the relevant code to the http-agent.js file and also updated
the unit tests for this module.
