+++ lib/request.js
@@ -1,12 +1,12 @@
 var
-	coreUtil = require('util'),
 	events = require('events'),
 	http = require('http'),
+	httpAgent = require('./http-agent'),
 	https = require('https'),
 	qs = require('querystring'),
 	url = require('url'),
-	utils = require('./utils'),
-	validation = require('./validation');
+	validation = require('./validation'),
+	util = require('util');
 
 const
 	DEFAULT_KEEP_ALIVE = true,
@@ -168,11 +168,10 @@ module.exports = (function (self) {
 						DEFAULT_KEEP_ALIVE);
 
 				// here we make sure that the appropriate http.Agent is used
-				if (augmented.keepAlive) {
-					augmented.agent = utils.getHTTPAgent.keepAlive(augmented.secure);
-				} else {
-					augmented.agent = utils.getHTTPAgent.default(augmented.secure);
-				}
+				augmented.agent = httpAgent.getAgent(
+					augmented.secure,
+					augmented.keepAlive
+				);
 
 				// create `path` from pathname and query.
 				augmented.path = validation.coalesce(
@@ -454,7 +453,7 @@ module.exports = (function (self) {
 	}
 
 	// enable events
-	coreUtil.inherits(Request, events.EventEmitter);
+	util.inherits(Request, events.EventEmitter);
 
 	// return the ability to create a new request
 	self.Request = Request;
