+++ package.json
@@ -37,6 +37,10 @@
     {
       "name": "Philippe Vachon-Rivard",
       "url": "https://github.com/privard"
+    },
+    {
+      "name": "Ian Calderon",
+      "url": "https://github.com/iancl"
     }
   ],
   "description": "SDK to make accessing PlayNetwork APIs when writing code in Javascript (Node.js) much easier.",
